v3 data audit: corrected TPU comparison + 50K results
- Paper: TPU table now uses parameter-matched baseline (317M vs 319M)
  with correct 200-step averaged PPL, extended to 50K steps
- Paper: Downstream evals use matched baseline (baseline wins at 20K)
- Paper: Reframed as convergence speed advantage (up to 43% faster)
  rather than absolute performance
- Proof Pack v3: All numbers verified from raw log files, corrections
  documented, honest about where baseline wins
- README: Added TPU convergence table, corrected claims
- Added TPU configs, eval scripts, setup scripts, chat server
- Added 1.57B gate ablation data file
- All 50K training runs complete (4/5; no-regime at 29K due to preemption)

+++ lolm/memory.py
@@ -61,7 +61,8 @@ class MemoryBank(nn.Module):
         # Attention over memory slots
         attn = torch.matmul(q, memory.transpose(-1, -2))  # (B, T, n_slots)
         attn = attn / (self.slot_dim ** 0.5)
-        attn = F.softmax(attn, dim=-1)
+        # Upcast to float32 for softmax to prevent bfloat16 NaN
+        attn = F.softmax(attn.float(), dim=-1).to(q.dtype)
         return torch.matmul(attn, memory), attn  # (B, T, slot_dim), (B, T, n_slots)
 
     def write(self, h: torch.Tensor, memory: torch.Tensor) -> torch.Tensor:
